+++ src/main/java/com/therapyCommunity_Vol1/backend/post/dto/SearchCursorResponse.java
@@ -3,6 +3,7 @@ package com.therapyCommunity_Vol1.backend.post.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -14,6 +15,10 @@ import java.util.List;
  * - data : 현재 페이지 게시글 요약 목록
  * - meta : hasNextData / nextScore / nextId
  *   nextScore, nextId 는 마지막 페이지일 경우 모두 null
+ *
+ * nextScore 가 BigDecimal 인 이유: pg_trgm similarity 결과는 PG 측에서 real(float4) 인데
+ * 클라이언트 왕복 시 부동소수 round-trip 정밀도 손실로 동등 비교(=)가 빗나갈 수 있다.
+ * Repository 쿼리에서 numeric(10,8) 로 캐스트한 값을 그대로 받아 BigDecimal 로 노출한다.
  */
 @Getter
 @AllArgsConstructor
@@ -26,7 +31,7 @@ public class SearchCursorResponse {
     @AllArgsConstructor
     public static class SearchCursorMeta {
         private final boolean hasNextData;
-        private final Double nextScore;
+        private final BigDecimal nextScore;
         private final Long nextId;
     }
 }
